Some shooting-from-the-hip code extensions, mainly centering around costume serialization. When I get hash_idx->part name string converter going i'll formalize the design.
git-svn-id: svn://svn.code.sf.net/p/segs/code2@123 a19be42f-918d-4297-8077-ab25d1e4f51f

+++ Components/Common/Bitstream/src/BitStream.cpp
@@ -287,7 +287,7 @@ s32 BitStream::uGetBits(u32 nBits)
 	tp = (u64 *)read_ptr();
 	r = *tp;//swap64(*tp);
 	r>>=m_read_bit_off; // starting at the top
-	tgt = r & (~1ull)>>(64-nBits);
+	tgt = s32(r & (~1ull)>>(64-nBits));
 	read_ptr((m_read_bit_off+nBits)>>3);
 	m_read_bit_off = (m_read_bit_off+nBits)&0x7; 
 	return tgt;
